fix: cast to int32 before left shift (>8) operators for 16bit systms
left shift first cast to int. So on 16 bit systems where int is 16bit
(like arduino) the result is expected to be wrong (and a warning is
shown at compilation)

+++ generator/C/include_v2.0/mavlink_helpers.h
@@ -703,7 +703,7 @@ MAVLINK_HELPER uint8_t mavlink_frame_char_buffer(mavlink_message_t* rxmsg,
 		break;
 
 	case MAVLINK_PARSE_STATE_GOT_MSGID2:
-		rxmsg->msgid |= c<<16;
+		rxmsg->msgid |= ((uint32_t)c)<<16;
 		mavlink_update_checksum(rxmsg, c);
 		if(rxmsg->len > 0){
 			status->parse_state = MAVLINK_PARSE_STATE_GOT_MSGID3;
